+++ announcements/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
 from announcements.models import Announcement
-from web.multilingual.database import MultiLingualFieldAdmin
+from web.multilingual.admin import MultiLingualFieldAdmin
 
 admin.site.register(Announcement, MultiLingualFieldAdmin)
