+++ card/modelfields.py
@@ -1,29 +1,32 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-import card.forms
+from .validators import card_number_validator
 
 
 class CardNumberField(models.CharField):
     """
     Custom field for card numbers, doing some extra validation
     """
+    default_validators = [card_number_validator]
 
     def __init__(self, **kwargs):
         kwargs = {
             "verbose_name": _("Card number"),
-            **kwargs,
-            "validators": card.forms.card_number_validators,
+            "validators": self.default_validators,
             "max_length": 10,
             "error_messages": {
                 "unique": _("Card number already in use"),
             },
+            **kwargs,
         }
         super().__init__(**kwargs)  # No card numbers are more than ten digits long
 
     def formfield(self, **kwargs):
+        from . import formfields  # avoids circular imports
+
         if "form_class" not in kwargs:
-            kwargs["form_class"] = card.forms.CardNumberField
+            kwargs["form_class"] = formfields.CardNumberField
         return super().formfield(**kwargs)
 
     def get_prep_value(self, value):
