+++ card/views.py
@@ -3,8 +3,8 @@ from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import csrf_exempt
 from django.views import View
 
-import card
 from web import settings
+from . import utils
 
 
 @method_decorator(csrf_exempt, name="dispatch")
@@ -26,7 +26,7 @@ class RFIDView(View):
             return HttpResponse(status=400)
 
         if secret == settings.CHECKIN_KEY:
-            if card.utils.is_valid(card_number):
+            if utils.is_valid(card_number):
                 return self.card_number_valid(card_number)
             else:
                 return self.card_number_invalid(card_number)
