+++ contentbox/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
 from contentbox.models import ContentBox
-from web.multilingual.database import MultiLingualFieldAdmin
+from web.multilingual.admin import MultiLingualFieldAdmin
 
 admin.site.register(ContentBox, MultiLingualFieldAdmin)
