+++ docs/models.py
@@ -1,17 +1,15 @@
 from ckeditor_uploader.fields import RichTextUploadingField
-from django.core.validators import RegexValidator
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 from users.models import User
+from .validators import page_title_validator
 
 
 class Page(models.Model):
     """Model for each individual documentation page"""
-    title_regex = r"^[0-9A-Za-z ():]+$"
-    title_validator = RegexValidator(regex=title_regex,
-                                     message=_("Only numbers, letters, space, parenthesises and colon are allowed"))
-    title = models.CharField(max_length=64, unique=True, verbose_name=_("Title"), validators=[title_validator])
+
+    title = models.CharField(max_length=64, unique=True, verbose_name=_("Title"), validators=[page_title_validator])
     created_by = models.ForeignKey(User, models.SET_NULL, null=True, blank=True)
     current_content = models.ForeignKey("Content", models.SET_NULL, null=True, blank=True, related_name="in_use")
