+++ make_queue/forms.py
@@ -4,12 +4,13 @@ from django.db.models import Q
 from django.forms import ModelChoiceField, IntegerField
 from django.utils.translation import gettext_lazy as _
 
-import card
-from card.forms import CardNumberField
+from card import utils as card_utils
+from card.formfields import CardNumberField
 from news.models import TimePlace
 from users.models import User
 from web.widgets import SemanticTimeInput, SemanticChoiceInput, SemanticSearchableChoiceInput, SemanticDateInput, \
     MazemapSearchInput
+from .formfields import UserModelChoiceField
 from .models.course import Printer3DCourse
 from .models.models import Machine, MachineType, Quota, ReservationRule
 
@@ -104,11 +105,6 @@ class RuleForm(forms.ModelForm):
 
 
 class QuotaForm(forms.ModelForm):
-    class UserModelChoiceField(ModelChoiceField):
-
-        def label_from_instance(self, obj):
-            return f'{obj.get_full_name()} - {obj.username}'
-
     user = UserModelChoiceField(queryset=User.objects.all(),
                                 widget=SemanticSearchableChoiceInput(prompt_text=_("Select user")),
                                 label=_("User"),
@@ -163,7 +159,7 @@ class Printer3DCourseForm(forms.ModelForm):
     def is_valid(self):
         card_number = self.data["card_number"]
         username = self.data["username"]
-        is_duplicate = card.utils.is_duplicate(card_number, username)
+        is_duplicate = card_utils.is_duplicate(card_number, username)
         if is_duplicate:
             self.add_error("card_number", _("Card number is already in use"))
         return super().is_valid() and not is_duplicate
